# Allocation free synthesis