--- a/src/bellpepper/solver.rs
+++ b/src/bellpepper/solver.rs
@@ -1,11 +1,11 @@
-use crate::traits::Group;
+use crate::traits::Engine;
 use bellpepper_core::{ConstraintSystem, Index, LinearCombination, SynthesisError, Variable};
 use ff::PrimeField;
 
 use bellpepper::util_cs::witness_cs::WitnessCS;
 
 /// A `ConstraintSystem` which calculates witness values for a concrete instance of an R1CS circuit.
-pub type SatisfyingAssignment<G> = WitnessCS<<G as Group>::Scalar>;
+pub type SatisfyingAssignment<E> = WitnessCS<<E as Engine>::Scalar>;
 
 #[derive(Debug, PartialEq, Eq)]
 /// A `ConstraintSystem` which calculates witness values for a concrete instance of an R1CS circuit.