Addresses #137.

This PR introduces `WitnessViewCS`, a slightly modified version of `WitnessCS` that does not own the witness buffers. Instead, we have `RecursiveSNARK` own the witness buffers, and `WitnessViewCS` takes mutable handles into those buffers. 

By allowing `RecursiveSNARK` to directly manage its own witness memory, we remove the extra allocations `WitnessCS` demands. 

Benchmark results from downstream `lurk-rs` confirm that this improves performance: https://github.com/lurk-lab/lurk-rs/pull/922

## Pasted Benchmark
  <details>
    <summary>Click to view benchmark</summary>

| Test | Base         | PR               | % |
|------|--------------|------------------|---|
| LEM Fibonacci Prove - rc = 100/fib/num-100 | 6.0±0.06s | **5.6±0.04s** | **-6.67%** |
| LEM Fibonacci Prove - rc = 100/fib/num-200 | 12.4±0.07s | **12.0±0.04s** | **-3.23%** |
| LEM Fibonacci Prove - rc = 600/fib/num-100 | 5.3±0.06s | N/A | N/A |
| LEM Fibonacci Prove - rc = 600/fib/num-200 | 12.1±0.04s | N/A | N/A |

  </details>