@winston-h-zhang 2023-11-27
# Benchmarks

## Table of Contents

- [Benchmark Results](#benchmark-results)
    - [LEM Fibonacci Prove Fresh - rc = 100](#lem-fibonacci-prove-fresh---rc-=-100)
    - [LEM Fibonacci Prove Fresh - rc = 600](#lem-fibonacci-prove-fresh---rc-=-600)
    - [LEM Fibonacci Prove Load - rc = 100](#lem-fibonacci-prove-load---rc-=-100)
    - [LEM Fibonacci Prove Load - rc = 600](#lem-fibonacci-prove-load---rc-=-600)

## Benchmark Results

### LEM Fibonacci Prove Fresh - rc = 100

|               | `main/fib`             | `witness-view-cs/fib`  | `new-witness-cs/fib`   |
|:--------------|:---------------------- |:---------------------- |:---------------------- |
| **`num-100`** | `6.39 s` (**1.00x**)   | `7.14 s` (**1.11x**)   | `6.91 s` (**1.00x**)   |
| **`num-200`** | `13.59 s` (**1.00x**)  | `14.26 s` (**1.05x**)  | `14.76 s` (**1.00x**)  |
| **`adthrpt`** | `222.2` (**1.00x**)    | `224.7`   (**1.01x**)  | `203.8`   (**0.92x**)  |

### LEM Fibonacci Prove Fresh - rc = 600

|               | `main/fib`             | `witness-view-cs/fib`  | `new-witness-cs/fib`   |
|:--------------|:---------------------- |:---------------------- |:---------------------- |
| **`num-100`** | `4.54 s` (**1.00x**)   | `4.46 s` (**0.98x**)   | `4.45 s` (**1.00x**)   |
| **`num-200`** | `10.58 s` (**1.00x**)  | `10.38 s` (**0.98x**)  | `10.40 s` (**1.00x**)  |
| **`adthrpt`** | `298.0` (**1.00x**)    | `304.0`   (**1.02x**)  | `302.5`   (**1.01x**)  |

### LEM Fibonacci Prove Load - rc = 100

|               | `main/fib`             | `witness-view-cs/fib`  | `new-witness-cs/fib`   |
|:--------------|:---------------------- |:---------------------- |:---------------------- |
| **`num-100`** | `6.39 s` (**1.00x**)   | `6.30 s` (**0.98x**)   | `6.26 s` (**1.00x**)   |
| **`num-200`** | `13.05 s` (**1.00x**)  | `12.68 s` (**0.97x**)  | `12.66 s` (**1.00x**)  |
| **`adthrpt`** | `240.2` (**1.00x**)    | `250.8`   (**1.04x**)  | `250.0`   (**1.04x**)  |

### LEM Fibonacci Prove Load - rc = 600

|               | `main/fib`             | `witness-view-cs/fib`  | `new-witness-cs/fib`   |
|:--------------|:---------------------- |:---------------------- |:---------------------- |
| **`num-100`** | `5.78 s` (**1.00x**)   | `5.74 s` (**1.00x**)   | `5.77 s` (**1.00x**)   |
| **`num-200`** | `15.25 s` (**1.00x**)  | `13.33 s` (**0.87x**)  | `13.59 s` (**1.00x**)  |
| **`adthrpt`** | `190.1` (**1.00x**)    | `237.2`   (**1.25x**)  | `230.2`   (**1.21x**)  |

---
Made with [criterion-table](https://github.com/nu11ptr/criterion-table)